requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via
    #   -r requirements.in
    #   django-celery-results
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==5.0.3
    # via
    #   -r requirements.in
    #   django-celery-results
    #   django-filter
    #   django-tenant-users
    #   django-tenants
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-results==2.5.1
    # via -r requirements.in
django-filter==24.1
    # via -r requirements.in
django-tenant-users==1.3.0
    # via -r requirements.in
django-tenants==3.6.1
    # via
    #   -r requirements.in
    #   django-tenant-users
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
kombu==5.3.5
    # via celery
prompt-toolkit==3.0.43
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
pytz==2024.1
    # via djangorestframework
redis==5.0.3
    # via celery
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via asgiref
tzdata==2024.1
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
